file_share/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
pillow==11.0.0
PyJWT==2.10.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
